up to 70%

+++ tools/coverage-runner.py
@@ -10,10 +10,31 @@ import os
 #--------------------------------------------------------------------------#
 # Helpers
 #--------------------------------------------------------------------------#
-def runner(module, output):
-    cmd = "nosetests --with-coverage --cover-package=%s" % module
-    os.system(cmd)
-    os.system("rm .coverage")
+def preRunner(options):
+    ''' Runs the pre testing commands
+    @param options The collection of options
+    '''
+    if options.coverage:
+        os.system("rm .coverage")
+
+def mainRunner(options):
+    ''' Runs the testing with the supplied options
+    @param options The collection of options
+    '''
+    cmd = "nosetests"
+    if options.coverage:
+        cmd += " --with-coverage --cover-package=%s" % options.module
+    if options.output != "":
+        cmd += " 2>&1 > %s" % options.output
+    if options.unittest:
+        os.system(cmd)
+
+def postRunner(options):
+    ''' Runs the post testing commands
+    @param options The collection of options
+    '''
+    if options.coverage:
+        os.system("rm .coverage")
 
 #--------------------------------------------------------------------------#
 # Main Runner
@@ -25,7 +46,13 @@ def main():
     parser = OptionParser()
     parser.add_option("-o", "--output",
                     help="Where to store coverage results",
-                    dest="output", default="coverage.results")
+                    dest="output", default="")
+    parser.add_option("-u", "--unittest",
+                    help="Run nose unit tests",
+                    action="store_true", dest="unittest", default=True)
+    parser.add_option("-c", "--coverage",
+                    help="Run code coverage",
+                    action="store_true", dest="coverage", default=False)
     parser.add_option("-m", "--module",
                     help="The module to run coverage tests on",
                     dest="module", default="nose")
@@ -33,7 +60,9 @@ def main():
         (opt, arg) = parser.parse_args()
 
         print "Running Code Coverage Of: %s\n" % opt.module
-        runner(opt.module, opt.output)
+        preRunner(opt)
+        mainRunner(opt)
+        postRunner(opt)
 
     except Exception, ex:
         print ex
